+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog of lizard-annotation
 0.5.7 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Let form pass extent to geometry editor.
 
 
 0.5.6 (2012-04-25)